Feature: Complex test IFileStorage impls
